Change column types to float

+++ Pizzeria.Domain/Dto/IngredientDto/RequestIngredientDto.cs
@@ -14,6 +14,6 @@ namespace Pizzeria.Domain.Dto.IngredientDto
         public required decimal IngredientPrice { get; set; }
 
         [Required]
-        public required long QuantityInStock { get; set; }
+        public required float QuantityInStock { get; set; }
     }
 }
